Override equals/hashCode for references.
References are used as keys in a hash map inside resolve cache, and
there may be several different GoReference objects for a single PSI
element, so providing proper equality semantics is necessary to make the
PSI cache most efficient.

So, the references are considered equal iff they share the same
underlying PSI element.